refactor: modularize modals and navigation
